requirements.txt
flask
openai
Pillow
numpy
